pip_requirements.txt
# absl-py==2.1.0
# accelerate==0.34.2
# addict==2.4.0
# antlr4-python3-runtime==4.9.3
# as-list==0.2.1
# asciitree==0.3.3
# astunparse==1.6.3
# attrs==24.2.0
# av==10.0.0
# bcrypt==4.2.0
# bitsandbytes==0.43.2
# blinker==1.8.2
# cachetools==5.4.0
# certifi==2024.7.4
# cffi==1.16.0
# charset-normalizer==3.3.2
# git+https://github.com/openai/CLIP.git
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
configargparse==1.7
contourpy==1.1.1
cryptography==43.0.1
cycler==0.12.1
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
deprecated==1.2.14
diffusers==0.30.2
docker-pycreds==0.4.0
easydict==1.13
egl-probe==1.0.2
einops==0.8.0
entrypoints==0.4
fabric==3.2.2
fasteners==0.19
fastjsonschema==2.20.0
filelock==3.15.4
fire==0.6.0
flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
freetype-py==2.4.0
fsspec==2024.6.1
ftfy==6.2.0
gast==0.4.0
gitdb==4.0.11
gitpython==3.1.43
google-auth==2.32.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.65.1
gym==0.21.0
h5py==3.11.0
huggingface-hub==0.24.0
hydra-core==1.3.2
icecream==2.1.3
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.4.7
importlib-metadata==8.0.0
importlib-resources==6.4.0
invoke==2.2.0
iopath==0.1.10
ipycanvas==0.13.2
ipywidgets==8.1.3
itsdangerous==2.2.0
jax==0.4.13
jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyterlab-widgets==3.0.11
keras==2.12.0
kiwisolver==1.4.5
libclang==18.1.1
markdown==3.6
markupsafe==2.1.5
matplotlib==3.7.5
ml-dtypes==0.2.0
mpmath==1.3.0
natsort==8.4.0
nbformat==5.10.4
networkx==3.1
numcodecs==0.12.1
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
open3d==0.18.0
opencv-python==4.10.0.84
opt-einsum==3.3.0
orjson==3.10.7
packaging==24.1
paramiko==3.5.0
peft==0.12.0
pexpect==4.9.0
pkgutil-resolve-name==1.3.10
platformdirs==4.2.2
plotly==5.24.1
portalocker==2.10.1
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pudb==2024.1.2
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
pygame==2.1.2
pyglet==2.0.16
pymunk==6.2.1
pynacl==1.5.0
pyopengl==3.1.0
pyparsing==3.1.2
pyquaternion==0.9.9
pyrender==0.1.45
pytz==2024.1
pywavelets==1.4.1
pyyaml==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
robomimic==0.2.0
rpds-py==0.20.0
rsa==4.9
runstats==2.0.0
safetensors==0.4.3
scikit-image==0.19.3
scikit-learn==1.3.2
scikit-video==1.1.11
scipy==1.10.1
sentry-sdk==2.12.0
setproctitle==1.3.3
shapely==1.8.4
six==1.16.0
smmap==5.0.1
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorboardx==2.6.2.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
thop==0.1.1-2209072238
threadpoolctl==3.1.0
tifffile==2023.7.10
timeout-decorator==0.5.0
timm==1.0.7
tokenizers==0.19.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.42.4
transforms3d==0.4.2
trimesh==4.4.3
triton==2.3.1
typing-extensions==4.12.2
urllib3==2.2.2
urwid==2.6.15
urwid-readline==0.14
wandb==0.17.5
wcwidth==0.2.13
werkzeug==3.0.3
widgetsnbextension==4.0.11
wrapt==1.14.1
yacs==0.1.8
zarr==2.16.1
zipp==3.19.2
gymnasium==0.29.1
datasets==3.0.1
numba==0.58.1
deepdiff==8.0.1
colormath==3.0.0